+++ nvblox/executables/include/nvblox/datasets/3dmatch.h
@@ -19,7 +19,7 @@ limitations under the License.
 #include <string>
 
 #include "nvblox/core/types.h"
-#include "nvblox/datasets/data_loader.h"
+#include "nvblox/datasets/data_loader_interface.h"
 #include "nvblox/datasets/image_loader.h"
 #include "nvblox/executables/fuser.h"
 
@@ -28,9 +28,9 @@ namespace datasets {
 namespace threedmatch {
 
 // Build a Fuser for the 3DMatch dataset
-std::unique_ptr<Fuser> createFuser(const std::string base_path,
-                                   const int seq_id,
-                                   bool init_from_gflags = true);
+std::unique_ptr<CameraFuser> createFuser(const std::string base_path,
+                                         const int seq_id,
+                                         bool init_from_gflags = true);
 
 ///@brief A class for loading 3DMatch data
 class DataLoader : public RgbdDataLoaderInterface {
@@ -51,6 +51,9 @@ class DataLoader : public RgbdDataLoaderInterface {
                                             const int seq_id,
                                             bool multithreaded = true);
 
+  /// 3DMatch datasets do not provide frame timestamps
+  bool provides_frame_timestamps() const override { return false; }
+
   /// Interface for a function that loads the next frames in a dataset
   /// This version of the function should be used when the color and depth
   /// camera are the same.
@@ -62,7 +65,7 @@ class DataLoader : public RgbdDataLoaderInterface {
   DataLoadResult loadNext(DepthImage* depth_frame_ptr,  // NOLINT
                           Transform* T_L_C_ptr,         // NOLINT
                           Camera* camera_ptr,           // NOLINT
-                          ColorImage* color_frame_ptr = nullptr) override;
+                          ColorImage* color_frame_ptr = nullptr);
 
   /// Interface for a function that loads the next frames in a dataset.
   /// This is the version of the function for different depth and color cameras.
@@ -72,13 +75,19 @@ class DataLoader : public RgbdDataLoaderInterface {
   ///@param[out] color_frame_ptr The loaded color frame.
   ///@param[out] T_L_C_ptr Transform from color camera to the Layer frame.
   ///@param[out] color_camera_ptr The intrinsic color camera model.
+  ///@param[out] unused Needed to match data loader interface (pass nullptr).
+  ///@param[out] unused Needed to match data loader interface (pass nullptr).
+  ///@param[out] unused Needed to match data loader interface (pass nullptr).
   ///@return Whether loading succeeded.
   DataLoadResult loadNext(DepthImage* depth_frame_ptr,  // NOLINT
                           Transform* T_L_D_ptr,         // NOLINT
                           Camera* depth_camera_ptr,     // NOLINT
                           ColorImage* color_frame_ptr,  // NOLINT
                           Transform* T_L_C_ptr,         // NOLINT
-                          Camera* color_camera_ptr) override;
+                          Camera* color_camera_ptr,     // NOLINT
+                          Time*,                        // NOLINT
+                          Transform*,                   // NOLINT
+                          Time*) override;              // NOLINT
 
  protected:
   const std::string base_path_;