+++ nvblox/executables/src/datasets/replica.cpp
@@ -142,13 +142,14 @@ std::unique_ptr<ImageLoader<ColorImage>> createColorImageLoader(
 
 }  // namespace internal
 
-std::unique_ptr<Fuser> createFuser(const std::string base_path,
-                                   bool init_from_gflags) {
+std::unique_ptr<CameraFuser> createFuser(const std::string base_path,
+                                         bool init_from_gflags) {
   auto data_loader = DataLoader::create(base_path, false);
   if (!data_loader) {
-    return std::unique_ptr<Fuser>();
+    return std::unique_ptr<CameraFuser>();
   }
-  return std::make_unique<Fuser>(std::move(data_loader), init_from_gflags);
+  return std::make_unique<CameraFuser>(std::move(data_loader),
+                                       init_from_gflags);
 }
 
 DataLoader::DataLoader(const std::string& base_path, bool multithreaded)
@@ -277,12 +278,10 @@ DataLoadResult DataLoader::loadNext(DepthImage* depth_frame_ptr,
   return DataLoadResult::kSuccess;
 }
 
-DataLoadResult DataLoader::loadNext(DepthImage* depth_frame_ptr,
-                                    Transform* T_L_D_ptr,
-                                    Camera* depth_camera_ptr,
-                                    ColorImage* color_frame_ptr,
-                                    Transform* T_L_C_ptr,
-                                    Camera* color_camera_ptr) {
+DataLoadResult DataLoader::loadNext(
+    DepthImage* depth_frame_ptr, Transform* T_L_D_ptr, Camera* depth_camera_ptr,
+    ColorImage* color_frame_ptr, Transform* T_L_C_ptr, Camera* color_camera_ptr,
+    Time*, Transform*, Time*) {
   // NOTE: The other pointers are checked non-null below
   CHECK_NOTNULL(color_frame_ptr);
   CHECK_NOTNULL(T_L_C_ptr);